package structure refactored and unused classes removed, #48

+++ gradoop-biiig/src/main/java/org/biiig/fsm/gspan/common/FrequentLabel.java
@@ -16,7 +16,7 @@ public class FrequentLabel implements Comparable<FrequentLabel> {
 
   @Override
   public int compareTo(FrequentLabel other) {
-    int comparison = (int)(other.support - this.support);
+    int comparison = other.support - this.support;
 
     if (comparison == 0) {
